requirements.txt
anyio==4.11.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
exceptiongroup==1.3.0
firebase_admin==7.1.0
Flask==3.1.2
google-api-core==2.28.1
google-auth==2.42.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgpack==1.1.2
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pynmea2==1.19.0
pyserial==3.5
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
shiboken6==6.10.0
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==1.26.20
Werkzeug==3.1.3
zipp==3.23.0
